requirements.txt
fastapi
uvicorn
python-multipart
jinja2
rdflib
pyshacl
fhir.resources
pandas
numpy
scipy
reportlab
pydantic
gunicorn